add styling to the form.
