+++ index.html
@@ -14,10 +14,10 @@
   <nav id="nav-id" class="navbar navbar-expand-lg navbar-info bg-light justify-content-between">
     <a class="navbar-brand" href="#">Library</a>
   </nav>
-  <div class="page-container">
+  <div id="page-container">
     <div id="content">
     </div>
   </div>
-  <script src="assets/index.js"></script>
+  <script type="module" src="assets/index.js"></script>
 </body>
 </html>
